refactor(optimizer): fold match_thread into labeled_block_fusion
Both rules eliminated the intermediate variant an inlined helper leaves
at a variant-discriminating consumer; they only differed in the shape
of that consumer and whether its value is observed. Merge them into one
rule with two entry points:

- apply_block: the value-discarding if-let / two-statement-match fusion
  (unchanged).
- apply_expr: the value-producing `match LB { … }` threading (the `?`
  shape), moved verbatim from the deleted match_thread module.

They already shared the loop-safety helpers and single_payload_binding;
co-locating them removes the separate pass and its registration. Output
is byte-identical — the golden WIR fixtures are unchanged.

Full e2e (O0/O2) 3248 and test-wado 3599 green; clippy clean.
