+++ docs/optimizer.md
@@ -58,7 +58,7 @@ Gating affects only which functions a pass visits, never the IR a visit produces
 `optimize.rs` orchestrates the NIR stages; `wir_optimize.rs` runs the WIR stages.
 
 1. Early DCE — remove unreachable functions/types/globals.
-2. Fixed-point loop (skipped at `-O0`), in order: `container_sroa`, peephole (pre-inline; hosts `match_to_switch` and `value_copy_elide` as rules), `value_copy_demote`, `sroa_param`, `inline`, peephole (post-inline; hosts `array_literal`, `labeled_block_fusion`, `match_thread`, `ref_elim`, and `elide_box_local` as rules), `sroa`, `copy_prop`, `dae`, `drve`, `const_folding`, `licm`, `tmpl_hoist`. (`match_to_switch` on global initializers runs once before the loop; `-O0` lowers everything via `match_to_switch_all`.)
+2. Fixed-point loop (skipped at `-O0`), in order: `container_sroa`, peephole (pre-inline; hosts `match_to_switch` and `value_copy_elide` as rules), `value_copy_demote`, `sroa_param`, `inline`, peephole (post-inline; hosts `array_literal`, `labeled_block_fusion`, `ref_elim`, and `elide_box_local` as rules), `sroa`, `copy_prop`, `dae`, `drve`, `const_folding`, `licm`, `tmpl_hoist`. (`match_to_switch` on global initializers runs once before the loop; `-O0` lowers everything via `match_to_switch_all`.)
 3. Post-loop, once: `field_scalarize`; `store_load_forward` (fold the scalarization shadow inits); `branch_prune_final` (flatten `__tmpl:` wrappers); `const_object_globalization` + a final `const_folding`/`const_branch_prune`/`condition_implication` cleanup.
 4. Final DCE.
 5. Backend-required rewrites (all levels): `select_lowering`, `multi_value_return`.
@@ -76,8 +76,7 @@ Allocation-elimination and aggregate passes:
 - `array_literal` — materialize the `array_new(N) + N×push` builder window into `ArrayLiteral` (lowered to `array.new_fixed`).
 - `value_copy_elide` — strip a `$value_copy$T(arg)` wrapper whose target is observably read-only.
 - `value_copy_demote` — demote a deep `List<E>` value-copy to a shallow spine copy when elements are provably never mutated through the binding (element-immutability taint analysis).
-- `labeled_block_fusion` — fuse the inlined-`Option<T>` `let __tmp = label:{… break Some(v)…}; if VariantTest(__tmp,Some)…` shape, deleting the variant allocation.
-- `match_thread` — jump-thread `match L:{… break L: Case(p) …} { arms }`, the value-producing shape `inline` leaves when a variant-returning callee is a match scrutinee (every inlined `x = f()?`). Each break site becomes its statically selected arm (payload bound, divergent arms keep their early exit), and the labeled block yields the match result directly — the intermediate variant is never allocated or dispatched on. Rewrites the `Match` node in place, so it applies in any position; exits nested in `Switch` / inner-`match` arms thread too. Bails on guards, non-`Variant`/`Wildcard` patterns, `null` (`Option::None`) breaks, and loop-exit capture.
+- `labeled_block_fusion` — one rule, two entry points, both deleting the intermediate variant an inlined `Option<T>`/`Result<T,E>` helper leaves at its consumer. `apply_block` fuses the value-discarding `let __tmp = label:{… break Some(v)…}; if VariantTest(__tmp,Some)…` if-let shape. `apply_expr` threads the value-producing `match L:{… break L: Case(p) …} { arms }` shape (the direct-scrutinee form every inlined `x = f()?` lowers to): each break site becomes its statically selected arm (payload bound, divergent arms keep their early exit) and the labeled block yields the match result directly, rewriting the `Match` node in place so it applies in any position (exits nested in `Switch` / inner-`match` arms thread too). The threading half bails on guards, non-`Variant`/`Wildcard` patterns, `null` (`Option::None`) breaks, and loop-exit capture.
 - `ref_elim` — drop reference bindings (`let r = &x`) read only via field access, rewriting reads to the original. Also collapses `let r = &<pure inline aggregate>` (a shared borrow of a `StructLiteral`, e.g. the inlined `len(&self)` binding `self = &String { … }`), substituting the aggregate at each `r.field` use so `const_folding`'s `project_struct_literal` folds the projection — the mechanism that makes `"x".len()` fold to a constant.
 
 Scalar / dataflow passes:
