+++ wado-compiler/src/optimize/arena_query.rs
@@ -11,10 +11,12 @@ use crate::nir_arena::{
 };
 
 /// The single optional payload-binding local of a variant arm pattern's
-/// `bindings`. `Some(None)` = no binding (`[]` or `[_]`); `Some(Some(idx))` =
-/// one `Binding` slot; `None` = reject (more than one binding, or a nested
-/// non-trivial subpattern the `labeled_block_fusion` / `match_thread` payload
-/// substitution does not handle). Callers propagate the reject with `?`.
+/// `bindings`. The two levels are distinct outcomes callers must tell apart, so
+/// `?` propagates the reject: `Some(None)` = no binding (`[]` or `[_]`);
+/// `Some(Some(idx))` = one `Binding` slot; `None` = reject (more than one
+/// binding, or a nested non-trivial subpattern the `labeled_block_fusion`
+/// value-discarding / value-producing payload substitution does not handle).
+#[allow(clippy::option_option)]
 pub(super) fn single_payload_binding(body: &Body, bindings: &[PatId]) -> Option<Option<u32>> {
     match bindings {
         [] => Some(None),
